feature(fs): Add Unicode aware readers to unify the text processing inside the engine
Following Unicode recommended pratices, this commit adds a unified interface that will try to convert anything that we encounter to UTF-8 encoded strings, since this is the standard encoding for the Go programming language. In this early version, it deals only with UTF-16 with Byte Order Marks (BOMs) so its simple, but already defines the standard encoding from now on, being an extensible part to support more corner cases in the future.