feature(engine): Add the primary interface for analysis
This commits adds the three main components of the engine, the Units, the Rules and the Findings. The first two acts as the inputs for anything the engine can do, so they are extensible and flexible interfaces to ensure a standardized way of finding code issues. This is just the ground work for sintatic analysis to happen within the engine, the next steps being define the default units to provide some base work for a user to run analysis. The Findings are just a agnostic way of reporting a finding, leaving all the work of documenting and actually reporting to a end-user to the user program.