feature(text): Add default TextUnit
With the base structure done, this commits adds the first concrete usable type for analysing text program, with convinience methods for dealing with UTF-16 encoding on Windows machines and also a concurrent way of dealing with rule per file. This asynchronous behaviour could be tricky to debug, so a TODO for the future is a solid reporting framework for errors from inside the engine.