fix(text): Fix concorrency issues with AND rules
Since our concurrency model depends on every goroutine returning a result, even if its empty, some issues in the code were leading to infinite runs, as the main routine (the one who sync all the results) kept waiting forever after all the sub routines were dead. This was fixed in several commits in the past, but this one, thanks to the weird logic for Regular, Or, And and Not rule types, just kept hiding from our attention. Hopefully, for now, this is done.