feat(Android): Add the first specialized Unit for AndroidManifest.xml files
To address the lack of depth in the Android analysis, this commits adds a ManifestUnit, that uses XPath 1.0 expressions to extend the precision of which we can go with the engine analyzing Android applications. It is important to remember, that even the if the Unit is called a ManifestUnit, on paper it can handle any XML document, so it can be renamed in the future to better suit the right semantics.