ir:feat - initial implementation of AST to IR conversion
This commit introduce a initial implementation of AST to IR conversion.

The new package IR declared two basic interfaces; Value and Instruction.
Value are IR nodes that holds a value that could be references by an
Instruction, like a variable or a constant value. Instruction on the
other hand are IR nodes that can compute a value or has some effect,
like binary expressions or call expressions.

The implementation is mostly based on golang.org/x/tools/go/ssa [1]
package.

The currently implementation only support call expressions and variable
declarations inside a function. The remaining conversions will be
implemented in the following commits with some test cases.

[1] https://pkg.go.dev/golang.org/x/tools@v0.1.8/go/ssa

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>