ir:fix - nil pointer on handle imports without name
This commit fix a panic on ir.NewFile that occur when an import don't
have a name, e.g: `import 'process'`. With this fix the imports.js tests
start to pass, so this commit also include the expected IR
representation.

A fix was also made on ir.WriteFile to print the imports in an orderly
fashion.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>