ir:feat - store source code information
This commit make possible to store source code information for IR nodes.
To make this possible a new struct node was created on ir package that
is embedded on all ir Value's and Instruction's that should provide
source code information. This struct implements the ast.Node interface
that provide the Pos method that returns the original source code
information.

Since the ast.Node interface is now implemented by third party packages
the node() method was removed.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>