ast:chore - don't panic on unsupported nodes
During the parsing of CST to AST, if some syntax is not supported yet
the default behaviour is to panic. Although it works well for the
development environment, allowing you to see bugs more quickly, to run
it experimentally in production can cause problems, as it makes it
impossible to analyze code that we support because it contains code that
we don't yet support.

This commit introduce a new node BadNode that represents nodes that
contains syntax errors or nodes that is not yet supported, with this we
can parse a entire Javascript file without any problemas and try to
analyze only syntaxes that we support.

Related with ZupIT/horusec#993

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>