ast:chore - store row position correctly directly
The lines returned by the tree-sitter always start at 0, to get around this
problem when we create an issue in analysis.NewIssue we add 1 more to the value
of the line, but implementing this in the analysis package makes the other calls
to node.Pos correct the value of the line.

This commit removes this need by correcting the row value during the creation of the
ast.Position object.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>