ir:feat - add support for global variable declarations
This commit add support to parse AST global variable declaration to IR
representation.

To add this support a new struct Global was created. This struct
implements the Value and Member interface and is added as a new file
member on File.Members field. The value of the global variable is not
translated during the parsing because could be hard to compute this
value, since we can don't have all file attributes parsed when parsing
global variable (for example, we can have a global variable declaration
that call a function declaration on the same file that was not parsed
yet) the value of global declaration is just an ast.Expr.

With this support the test of javascript/var_declarations.js start to
pass, so this commit also contains the expected IR representation to
this test.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>