ir:feat - add support to var declaration using call expressions
This commit adds support for translating AST-based variable declaration
using a call expression as its values to IR representation.

To add this support a new method String() was added on Value interface
to return the "disassembled" form of the Value if the Value is an
Instruction. This is useful to get an human-readable string
representation of a call expression. So, now Call node implements both
Value and Instruction interface.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>