ir:feat - add support for binary expressions
This commit introduce support for translating AST-based binary
expressions to IR representation.

The change introduced in ec049eb caused Call to implement the Value and
Instruction interfaces. Thus, when printing the Call expression, the
WriteFunction function called both String() and Name(), which for the
implementation of Call returned the same value, making the output have
the content: console.log('bug') = console.log('bug')
To fix this problem, the implementation of Call.Name() was changed to
return an empty string (a function call has no name at all) and a
validation was added in WriteFunction to print the name only if it is
not empty.

This commit also resolves a TODO in the Javascript CST to AST parse that
did not contain the binary expression operator. Since this operator was
added the expected test cases of AST parsing was updated too.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>