ir:feat - add support for template strings
This commit add support for translating AST-based template strings to IR
representation.

To add this support a new Value Template was created that holds a raw
string template and values that should be replaced.

With this support the javascript/functions.js tests start to pass. This
test was also changed to cover more scenarios of AST to IR translation.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>