ir:chore - avoid panic when parsing unsupported nodes
This commit introduce a new mechanism to be possible to avoid panic when
parsing an AST node that is not supported yet on IR.

Previously if some AST node is not supported on IR representation a
panic was generated, which is good for development because we can see
which syntaxes we don't support since the code sample used to test AST
parsing will fail to test IR parsing, but with this approach its hard to
enable an experimental semantic analysis on horusec cli because any
analysis that a user execute can cause a panic.

To fix this issue a new global variable function unsupportedNode was
created that could have two implementations;
    - Panic when debug mode is enable.
    - Do nothing when debug mode is disable.
The debug mode variable is changed at compile time and the default value
is disable (which will make it so that we don't need to change any cli
build configuration).

Related with ZupIT/horusec#993

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>