ir:chore - change Function.addNamedLocal to receive a Value
This commit change the signature of Function.addNamedLocal method to
receive a ir Value directly instead of parsing an raw expression. With
this we remove duplicate code from Function.addLocal. The method now
also expect a syntax parameter that holds the entire variable
declaration statement instead of just the identifier.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>