analysis/call:feat - analyze variable values
Previously if a variable declaration has a function call as a value this
function call was not being analysed by call.Analyzer which can generate
a lot of false positives, because it's common to have function calls
assigned to variables.

This commit improve the call analyzer check to inspect variable values
that could be instructions, so if a variable declaration has a function
call as its value, this function call will also be analyzed.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>