javascript:chore - add file information when assertNodeType fails
This commit add the file name, line and row on the panic message when
assertNodeType fails to assert that a node has a given type.

Since the assertNodeType now include the file name that is fails to
assert, this function was change to be a method of parser struct that
now holds the file name.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>