ir:feat - initial implementation of TAC based IR
This commit change the translation of AST to IR to use a TAC(Three
Addres Code)[1] intermediate representation.

This new representation simplify how we represent some complex codes.
Binary expressions that have more than two operands are converted to
multiple statements using temporary variables, for example, the binary
expression y = 10 + 20 * (10 / x) generate the following IR:
%t0 = 10 / x
%t1 = 20 * %t0
y = 10 + %t1

Call expressions was also simplified; if an argument of a call
expression is also a call expression e.g: foo(bar()) the following IR is
generated:
%t0 = bar()
foo(%t0)

With theses changes will be more easily to write rules, since we don't
need to handle multiple scenarios where we can have call expressions.
For example, we can have a call expression inside a if statement, or a
return statement, for both scenarios the IR generated will point these
call expressions to a temporary variable and use this temporary variable
created on the if statement and return statement, with this, a rule can
only check for call expressions when declaration a variable, instead to
check inside a if statement or a return statement.

To implement this, some constructs declared on ir/create.go was moved to
ir/builder.go and also changed to be a method from builder struct, with
this we can recursively call builder.expr that can expand complex
instructions to temporary variables.

The javascript/functions.js testcase was also changed to catch more
complex code during the parse to IR.

[1] https://en.wikipedia.org/wiki/Three-address_code

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>