ir:feat - add support for if statements
This commit add support to represent if statements on IR. This
representation is using Control-flow graph[1].

To add this support some new structs was created:
- If: Represents the if statement instruction. Contains the condition
  value of if and a basic block.
- Jump: Represents a jump instruction. This jump is used to represents
  jumps on if statement, for example, a jump from if them to if done or
  if else to if done. This jump instruction will also be used in other
  instructions that are represented using CFG's.

This commit add the ir/statements.js.out file that contains the if
statements IR code generated. Since we don't support all statements from
this test yet, this file was rewrited  using the debug flag disabled, so
the statements that we don't support is represented as empty blocks.
Running this test with debug enabled will fail, so we can still catch
not implemented translations from AST to IR.

[1] https://en.wikipedia.org/wiki/Control-flow_graph

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>