javascript:fix - not parsing multiple stmt's on else branch
Previously the case branch for StatementBlock was only parsing the first
child node of the block, which was causing wrong generated AST because
some nodes of else branch was missing.

This commit fix this issue by parsing all child nodes on StatementBlock
node correctly. A test case was also created to catch this bug.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>