ir:feature - adding class declaration instruction
Previously there was no treatment for classes, so attributes and methods
within these classes were not analyzed. This pull request adds support
for classes in the IR, where every class is represented as a struct .
For all methods of these structs a parameter was added with an
identifier that is a receiver.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>