javascript/ast:feat - add support to #nosec directives
This commit introduce the support to ignore block of codes using #nosec
comments.

The nosec directive can be used in any code block, the only requirement is
that the comment must be on the top line of the code that should be ignored,
comments on the same or below or more than one line above are not supported.
To view supported syntaxes see test file
internal/testdata/source/javascript/comments.js.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>