ir:feature - handling rhs identifiers
In some scenarios the value of rhs can be another identifier, this
scenario was not dealt previously in the IR. Now when there is a local
or global variable that uses another variable as a value, the original value of
the variable is used. For parameters we just set the value to nil.

Signed-off-by: Nathan Martins <nathan.martins@zup.com.br>