ir:feat - improve support of ast.SelectorExpr
Previously we was supporting ast.SelectorExpr only if the Expr field was
a identifier, so this commit improve this support handling the Expr
field recursively and resolving identifier imports.

Now we support expressions like `a.b.c()` and `a.b().c.d()`. For
selector expressions that Expr could be a call expression, temp
variables will be created and these temp variable will be used on the
name of selector expression.

Example:
    const a = a.b(10).c.d(20)
IR:
    %t0 = a.b(10)
    %t1 = %t0.c.d(20)

This apply not only for call expression but any expression on
ast.SelectorExpr.Expr that is not an *ast.Ident or *ast.SelectorExpr.

The *ast.SelectorExpr is also resolved when assigning variable values, but
we still need to improve this. When assigning a variable with a selector
expression, a new variable will be created to this selector expression
with the entire selector expression as a name of variable, but when
printing the IR, the value of this variable will be nil, because the
variable created from the selector expression has a nil value. This is
referenced in a TODO on *ast.SelectorExpr case from builder.expr method.

Lastly, the *ast.SelectorExpr is also resolved when the LHS operand of
an assign expression is an *ast.SelectorExpr. We still can improve this
support in the future, for example, when assigning a field from a class
e.g `this.a = 10`, we should be able to parse the `this` expression as
a receiver parameter from ir.Function object.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>