ast:fix - don't store comment nodes on AST
The commit e090c22 introduce a new Comment node that was being added on
AST. But adding this node on AST makes us need to handle them during parse
from AST to IR. This commit remove this Comment node, to avoid deal with
them on IR parsing.

Basically, when we don't need to check comments for #nosec we iterate
over childs nodes ignoring all nodes that are comments. When we need to
check for #nosec, we check and them ignore the node comment, so the
final AST will not contain any comment node.

The commit e090c22 also don't add the IR expected files for this tests,
so this commit also add these expected files.

Since the comment nodes are ignored, the AST expected test files was
also updated to remove comment nodes.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>