ir:feat - SSA based IR
This commit introduce a SSA based IR to semantic engine. The SSA
construction algorithm is based on the algorithm described on paper
Simple and Efficient Construction of Static Single Assignment Form.

Basically when parsing a variable declaration on AST, we store this
variable on current basic block of the function that the variable was
defined, when we need to lookup this variable we first check if the
current block contains the declared variable, if not exists we
recursively search on the predecessors blocks of the current basic block
and create a phi node with all possible values that this variable can
have.

The paper describe algorithms for optimizations but this commit only
implement the basic idea of store variables on basic block that defines
it and how we create phi nodes for variables that are defined in previous
basic blocks.

A new method lookup on builder struct was also created to lookup a
variable declaration inside a function level, function parameter level
or global level, this function is used to lookup a identifier instead of
created a new *Var to every *ast.Ident parsed. We only create a new *Var
if the builder.lookup return a nil value. So with this identifier on
binary expression, call expression and any other instruction will always
point to the correctly already defined ir.Value.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>