ir:feat - add support for for loop's
This commit add support for for loop's on IR.

Initially to add this support some bugs was fixed on parse from
tree-sitter CST to generic AST; Previously we was parsing the for
statement condition as a statement, which always result in a
ast.ExprStmt, so the Cond was changed to store the expression directly
which facilitated the parse in the IR. The increment node on for loop
could be nil in some syntaxes, or a empty_statement, so this commit also
handle these cases. A new field Op was also added on ast.IncExpr to
store the operator of a increment expression (++ or --).

On IR this commit add support for for loop's and increment expression
(since increment expressions is almost always used on for loop's). The
increment expression is converted to a binary expression, so a++ is
converted to a = a + 1. No new struct was needed to support for loops,
the implementation basically uses basic blocks and jumps to represent
the loops. With this loop representation a bug was also founded on
Function.lookup, since we can have recursively predecessors blocks that
points to each other the Function.recursivelyLookup it was going into an
infinite loop. To fix this issue, a map with the visited basic blocks
was created, so Function.recursivelyLookup check if the block being
visited was already visited, if yes stop the recursion, if no, search for
the local variable and add the basic block on the map of visited blocks.

This commit looks big, but the changes are almost all changes to the
test file.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>