javascript/ast:fix - new expression outside var decl
Previously the new_expression handle parser was always expecting that
the parent node was a variable declaration, witch it's not true because
we can have a code like this: foo(new Bar()).

This commit fix this issue by setting the ast.ObjectExpr.Name field only
if the parent node of new_expression is a variable_declarator.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>