ir:feat - add support for while statements
This commit add support to while statements on IR. This support follow
the same idea of for loop, we have a while.body, while.cond and
while.done blocks, the while.cond decide where to jump according to
condition, if true goto while.body, otherwise goto while.done,
while.body always jump to while.cond at the end to represent the loop.

Note: This implementation is not fully complete, in the future we should
have a better support for incomplete basic blocks (in this case the
while.cond is a incomplete block because the while.body is added as a
predecessor after the code on while.cod is emitted). This implementation
works well for what we need now, we just won't be able to correctly represent
a variable mutation inside the while.body that is used as a condition in
while.cond. In the future we should look at the 2.3 Handling Incomplete CFGs
section of the paper
"Simple and Efficient Construction of Static Single Assignment Form" to have
this implementation entirely correct.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>